// types/4_inline_query_result_voice.ts
import { MessageEntity } from "./0_message_entity.ts";
import { ParseMode } from "./0_parse_mode.ts";
import { InputMessageContent } from "./2_input_message_content.ts";
import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";

export interface InlineQueryResultVoice {
  type: "voice";
  id: string;
  voiceUrl: string;
  title: string;
  caption?: string;
  parseMode?: ParseMode;
  captionEntities?: MessageEntity[];
  voiceDuration?: number;
  replyMarkup?: InlineKeyboardMarkup;
  inputMessageContent?: InputMessageContent;
}